refactor: extract _SettingsLoader into dedicated module
