+++ connectors-sdk/tests/test_settings/test_settings_loader.py
@@ -3,10 +3,10 @@ from pathlib import Path
 from unittest.mock import patch
 
 import pytest
+from connectors_sdk.settings._settings_loader import _SettingsLoader
 from connectors_sdk.settings.base_settings import (
     BaseConfigModel,
     BaseConnectorSettings,
-    _SettingsLoader,
 )
 from connectors_sdk.settings.deprecations import DeprecatedField
 from pydantic import Field
